Split out `local-fs-store.hh`
This matches the already-existing `local-fs-store.cc`.
